+++ classes/class-ed-lti.php
@@ -113,6 +113,7 @@ class Ed_LTI {
 			$user_roles = new User_LTI_Roles( $tool->user->roles );
 			$blog_handler->add_user_to_blog( $user, $blog_id, $user_roles );
 			$blog_handler->add_user_to_top_level_blog( $user );
+			$blog_handler->set_additional_blog_options( $blog_handler->get_options_from_request( $_REQUEST ) );
 
 			$this->signin_user( $user, $blog_id );
 		}