+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] - 2020-06-28
+### Added
+- The plugin will now use a custom notification email supplied by the upstream VLE to store as a WordPress option for student blogs
 
 ## [1.2.1] - 2020-04-30
 ### Fixed